@@ -19,13 +19,14 @@
         ],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
-        "illuminate/database": "^6.0|^7.0|^8.0",
-        "illuminate/support": "^6.0|^7.0|^8.0"
+        "php": "^8.0",
+        "illuminate/database": "^7.0|^8.0",
+        "illuminate/support": "^7.0|^8.0",
+        "spatie/laravel-package-tools": "^1.6"
     },
     "require-dev": {
         "phpunit/phpunit" : "^8.0|^9.0",
-        "orchestra/testbench": "^4.0|^5.0|^6.0"
+        "orchestra/testbench": "^5.0|^6.0"
     },
     "autoload": {
         "psr-4": {
