@@ -17,9 +17,9 @@ class DeviceRecordUpdateSSHAction extends RowAction
     /**
      * 渲染模态框.
      *
-     * @return Modal|string
+     * @return Modal
      */
-    public function render()
+    public function render(): Modal
     {
         // 实例化表单类并传递自定义参数
         $form = DeviceSSHInfoForm::make()->payload(['id' => $this->getKey()]);
