@@ -17,9 +17,9 @@ class DeviceTrackUpdateDeleteAction extends RowAction
     /**
      * 渲染模态框.
      *
-     * @return Modal|string
+     * @return Modal
      */
-    public function render()
+    public function render(): Modal
     {
         $form = DeviceTrackUpdateDeleteForm::make()->payload([
             'id' => $this->getKey(),
