@@ -17,9 +17,9 @@ class TodoRecordUpdateAction extends RowAction
     /**
      * 渲染模态框.
      *
-     * @return Modal|string
+     * @return Modal
      */
-    public function render()
+    public function render(): Modal
     {
         $form = TodoRecordUpdateForm::make()->payload(['id' => $this->getKey()]);
 
