@@ -17,9 +17,9 @@ class ConsumableInAction extends AbstractTool
     /**
      * 渲染模态框.
      *
-     * @return Modal|string
+     * @return Modal
      */
-    public function render()
+    public function render(): Modal
     {
         return Modal::make()
             ->lg()
