@@ -10,6 +10,7 @@ use Exception;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Schema;
+use JetBrains\PhpStorm\ArrayShape;
 
 class CustomColumnDeleteAction extends RowAction
 {
@@ -32,7 +33,6 @@ class CustomColumnDeleteAction extends RowAction
      */
     public function handle(Request $request): Response
     {
-//        DB::beginTransaction();
         try {
             $table_name = $request->table_name;
             $name = $request->name;
@@ -56,12 +56,10 @@ class CustomColumnDeleteAction extends RowAction
                 $column_sort->delete();
             }
             $custom_column->delete();
-//            DB::commit();
             return $this->response()
                 ->success(trans('main.success'))
                 ->refresh();
         } catch (Exception $exception) {
-//            DB::rollBack();
             return $this->response()
                 ->error(trans('main.fail') . '：' . $exception->getMessage());
         }
@@ -77,6 +75,7 @@ class CustomColumnDeleteAction extends RowAction
         return [admin_trans_label('Delete Confirm'), admin_trans_label('Delete Confirm Description')];
     }
 
+    #[ArrayShape(['table_name' => "null|string", 'name' => "null|string"])]
     protected function parameters(): array
     {
         return [
