@@ -23,7 +23,7 @@ class ConsumableCategoryController extends AdminController
      * @param Request $request
      * @return mixed
      */
-    public function selectList(Request $request)
+    public function selectList(Request $request): mixed
     {
         $q = $request->get('q');
 
