@@ -91,7 +91,7 @@ class DeviceCategoryController extends AdminController
      * @param Request $request
      * @return mixed
      */
-    public function selectList(Request $request)
+    public function selectList(Request $request): mixed
     {
         $q = $request->get('q');
 
