@@ -41,7 +41,7 @@ class RoleController extends BaseRoleController
      * @param Request $request
      * @return mixed
      */
-    public function selectList(Request $request)
+    public function selectList(Request $request): mixed
     {
         $q = $request->get('q');
 
