@@ -100,11 +100,11 @@ class ServiceIssueController extends AdminController
     /**
      * Make a show builder.
      *
-     * @param mixed $id
+     * @param int $id
      *
      * @return Show
      */
-    protected function detail($id): Show
+    protected function detail(int $id): Show
     {
         return Show::make($id, new ServiceIssue(['service']), function (Show $show) {
             $show->field('id');
