@@ -24,7 +24,7 @@ class SoftwareCategoryController extends AdminController
      * @param Request $request
      * @return mixed
      */
-    public function selectList(Request $request)
+    public function selectList(Request $request): mixed
     {
         $q = $request->get('q');
 
