@@ -7,6 +7,7 @@ use Dcat\Admin\Layout\Column;
 use Dcat\Admin\Layout\Content;
 use Dcat\Admin\Layout\Row;
 use Dcat\Admin\Widgets\Card;
+use Illuminate\Contracts\Translation\Translator;
 
 class ToolChemexAppController extends Controller
 {
@@ -29,7 +30,7 @@ class ToolChemexAppController extends Controller
             });
     }
 
-    public function title()
+    public function title(): array|string|Translator|null
     {
         return admin_trans_label('title');
     }
