@@ -11,6 +11,7 @@ use Dcat\Admin\Grid;
 use Dcat\Admin\Http\Controllers\AdminController;
 use Dcat\Admin\Layout\Content;
 use Dcat\Admin\Show;
+use Illuminate\Contracts\Translation\Translator;
 use Illuminate\Http\Request;
 
 class VendorRecordController extends AdminController
@@ -30,7 +31,7 @@ class VendorRecordController extends AdminController
             ->body($this->grid());
     }
 
-    public function title()
+    public function title(): array|string|Translator|null
     {
         return admin_trans_label('title');
     }
@@ -98,7 +99,7 @@ class VendorRecordController extends AdminController
      *
      * @return mixed
      */
-    public function selectList(Request $request)
+    public function selectList(Request $request): mixed
     {
         $q = $request->get('q');
 
@@ -109,11 +110,11 @@ class VendorRecordController extends AdminController
     /**
      * Make a show builder.
      *
-     * @param mixed $id
+     * @param int $id
      *
      * @return Show
      */
-    protected function detail($id): Show
+    protected function detail(int $id): Show
     {
         return Show::make($id, new VendorRecord(), function (Show $show) {
             $show->field('id');
