@@ -2,226 +2,226 @@
 
 return [
     'scaffold' => [
-        'header' => '代码生成器',
-        'choose' => '选择已有数据表',
-        'table' => '表名',
-        'model' => '模型',
-        'controller' => '控制器',
-        'repository' => '数据仓库',
-        'add_field' => '添加字段',
-        'pk' => '主键',
-        'soft_delete' => '软删除',
-        'create_migration' => '创建表迁移文件',
-        'create_model' => '创建模型',
+        'header'            => '代码生成器',
+        'choose'            => '选择已有数据表',
+        'table'             => '表名',
+        'model'             => '模型',
+        'controller'        => '控制器',
+        'repository'        => '数据仓库',
+        'add_field'         => '添加字段',
+        'pk'                => '主键',
+        'soft_delete'       => '软删除',
+        'create_migration'  => '创建表迁移文件',
+        'create_model'      => '创建模型',
         'create_repository' => '创建数据仓库',
         'create_controller' => '创建控制器',
-        'run_migrate' => '创建数据表',
-        'create_lang' => '创建翻译文件',
-        'field' => '字段',
-        'translation' => '翻译',
-        'comment' => '注释',
-        'default' => '默认值',
-        'field_name' => '字段名',
-        'type' => '类型',
-        'nullable' => '允许空值',
-        'key' => '索引',
-        'translate_title' => '翻译标题',
+        'run_migrate'       => '创建数据表',
+        'create_lang'       => '创建翻译文件',
+        'field'             => '字段',
+        'translation'       => '翻译',
+        'comment'           => '注释',
+        'default'           => '默认值',
+        'field_name'        => '字段名',
+        'type'              => '类型',
+        'nullable'          => '允许空值',
+        'key'               => '索引',
+        'translate_title'   => '翻译标题',
         'sync_translation_with_comment' => '同步翻译与注释',
     ],
     'client' => [
-        'delete_confirm' => '确认删除?',
-        'confirm' => '确认',
-        'cancel' => '取消',
+        'delete_confirm'    => '确认删除?',
+        'confirm'           => '确认',
+        'cancel'            => '取消',
         'refresh_succeeded' => '刷新成功 !',
-        'submit' => '提交',
-        'close' => '关闭',
-        'selected_options' => '已选中:num个选项',
-        'exceed_max_item' => '已超出最大可选数量',
-        'no_preview' => '预览失败',
+        'submit'            => '提交',
+        'close'             => '关闭',
+        'selected_options'  => '已选中:num个选项',
+        'exceed_max_item'   => '已超出最大可选数量',
+        'no_preview'        => '预览失败',
 
         '500' => '系统繁忙，请稍后再试！',
         '403' => '对不起，您没有权限访问，请与管理员联系。',
         '401' => '请登录！',
         '419' => '对不起，当前页面已失效，请刷新浏览器。',
     ],
-    'home' => '主页',
-    'online' => '在线',
-    'login' => '登录',
-    'logout' => '登出',
-    'setting' => '设置',
-    'name' => '名称',
-    'username' => '用户名',
-    'old_password' => '旧密码',
-    'password' => '密码',
+    'home'                  => '主页',
+    'online'                => '在线',
+    'login'                 => '登录',
+    'logout'                => '登出',
+    'setting'               => '设置',
+    'name'                  => '名称',
+    'username'              => '用户名',
+    'old_password'          => '旧密码',
+    'password'              => '密码',
     'password_confirmation' => '确认密码',
-    'old_password_error' => '请输入正确的旧密码',
-    'remember_me' => '记住我',
-    'user_setting' => '用户设置',
-    'avatar' => '头像',
-    'list' => '列表',
-    'new' => '新增',
-    'create' => '创建',
-    'delete' => '删除',
-    'remove' => '移除',
-    'edit' => '编辑',
-    'quick_edit' => '快捷编辑',
-    'continue_editing' => '继续编辑',
-    'continue_creating' => '继续创建',
-    'view' => '查看',
-    'detail' => '详细',
-    'browse' => '浏览',
-    'reset' => '重置',
-    'export' => '导出',
-    'batch_delete' => '批量删除',
-    'save' => '保存',
-    'refresh' => '刷新',
-    'order' => '排序',
-    'expand' => '展开',
-    'collapse' => '收起',
-    'filter' => '筛选',
-    'search' => '搜索',
-    'close' => '关闭',
-    'show' => '显示',
-    'entries' => '条',
-    'captcha' => '验证码',
-    'action' => '操作',
-    'title' => '标题',
-    'description' => '简介',
-    'back' => '返回',
-    'back_to_list' => '返回列表',
-    'submit' => '提交',
-    'menu' => '菜单',
-    'input' => '输入',
-    'succeeded' => '成功',
-    'failed' => '失败',
-    'delete_confirm' => '确认删除?',
-    'delete_succeeded' => '删除成功 !',
-    'delete_failed' => '删除失败 !',
-    'update_succeeded' => '更新成功 !',
-    'update_failed' => '更新失败 !',
-    'save_succeeded' => '保存成功 !',
-    'save_failed' => '保存失败 !',
-    'refresh_succeeded' => '刷新成功 !',
-    'login_successful' => '登录成功 !',
-    'choose' => '选择',
-    'choose_file' => '选择文件',
-    'choose_image' => '选择图片',
-    'more' => '更多',
-    'deny' => '无权访问',
-    'administrator' => '管理员',
-    'roles' => '角色',
-    'permissions' => '权限',
-    'slug' => '标识',
-    'created_at' => '创建时间',
-    'updated_at' => '更新时间',
-    'alert' => '注意',
-    'parent_id' => '父级',
-    'icon' => '图标',
-    'uri' => '路径',
-    'operation_log' => '操作日志',
-    'parent_select_error' => '父级选择错误',
-    'default' => '默认',
-    'table' => '表格',
-    'no_data' => '暂无数据',
-    'routes' => '路由',
-    'alias' => '别名',
-    'route_action' => '处理器',
-    'middleware' => '中间件',
-    'import' => '导入',
-    'is_not_import' => '未导入',
-    'selected_options' => '已选中:num个选项',
-    'method' => '方法',
-    'user' => '用户',
-    'pagination' => [
+    'old_password_error'    => '请输入正确的旧密码',
+    'remember_me'           => '记住我',
+    'user_setting'          => '用户设置',
+    'avatar'                => '头像',
+    'list'                  => '列表',
+    'new'                   => '新增',
+    'create'                => '创建',
+    'delete'                => '删除',
+    'remove'                => '移除',
+    'edit'                  => '编辑',
+    'quick_edit'            => '快捷编辑',
+    'continue_editing'      => '继续编辑',
+    'continue_creating'     => '继续创建',
+    'view'                  => '查看',
+    'detail'                => '详细',
+    'browse'                => '浏览',
+    'reset'                 => '重置',
+    'export'                => '导出',
+    'batch_delete'          => '批量删除',
+    'save'                  => '保存',
+    'refresh'               => '刷新',
+    'order'                 => '排序',
+    'expand'                => '展开',
+    'collapse'              => '收起',
+    'filter'                => '筛选',
+    'search'                => '搜索',
+    'close'                 => '关闭',
+    'show'                  => '显示',
+    'entries'               => '条',
+    'captcha'               => '验证码',
+    'action'                => '操作',
+    'title'                 => '标题',
+    'description'           => '简介',
+    'back'                  => '返回',
+    'back_to_list'          => '返回列表',
+    'submit'                => '提交',
+    'menu'                  => '菜单',
+    'input'                 => '输入',
+    'succeeded'             => '成功',
+    'failed'                => '失败',
+    'delete_confirm'        => '确认删除?',
+    'delete_succeeded'      => '删除成功 !',
+    'delete_failed'         => '删除失败 !',
+    'update_succeeded'      => '更新成功 !',
+    'update_failed'         => '更新失败 !',
+    'save_succeeded'        => '保存成功 !',
+    'save_failed'           => '保存失败 !',
+    'refresh_succeeded'     => '刷新成功 !',
+    'login_successful'      => '登录成功 !',
+    'choose'                => '选择',
+    'choose_file'           => '选择文件',
+    'choose_image'          => '选择图片',
+    'more'                  => '更多',
+    'deny'                  => '无权访问',
+    'administrator'         => '管理员',
+    'roles'                 => '角色',
+    'permissions'           => '权限',
+    'slug'                  => '标识',
+    'created_at'            => '创建时间',
+    'updated_at'            => '更新时间',
+    'alert'                 => '注意',
+    'parent_id'             => '父级',
+    'icon'                  => '图标',
+    'uri'                   => '路径',
+    'operation_log'         => '操作日志',
+    'parent_select_error'   => '父级选择错误',
+    'default'               => '默认',
+    'table'                 => '表格',
+    'no_data'               => '暂无数据',
+    'routes'                => '路由',
+    'alias'                 => '别名',
+    'route_action'          => '处理器',
+    'middleware'            => '中间件',
+    'import'                => '导入',
+    'is_not_import'         => '未导入',
+    'selected_options'      => '已选中:num个选项',
+    'method'                => '方法',
+    'user'                  => '用户',
+    'pagination'            => [
         'range' => '从 :first 到 :last ，总共 :total 条',
     ],
-    'role' => '角色',
-    'permission' => '权限',
-    'route' => '路由',
-    'confirm' => '确认',
-    'cancel' => '取消',
-    'selectall' => '全选',
-    'http' => [
+    'role'                  => '角色',
+    'permission'            => '权限',
+    'route'                 => '路由',
+    'confirm'               => '确认',
+    'cancel'                => '取消',
+    'selectall'             => '全选',
+    'http'                  => [
         'method' => 'HTTP方法',
-        'path' => 'HTTP路径',
+        'path'   => 'HTTP路径',
     ],
-    'all_methods_if_empty' => '为空默认为所有方法',
-    'all' => '全部',
-    'current_page' => '当前页',
-    'selected_rows' => '选择的行',
-    'upload' => '上传',
-    'new_folder' => '新建文件夹',
-    'time' => '时间',
-    'size' => '大小',
-    'between_start' => '起始',
-    'between_end' => '结束',
-    'next_page' => '下一页',
-    'prev_page' => '上一页',
-    'next_step' => '下一步',
-    'prev_step' => '上一步',
-    'done' => '完成',
-    'listbox' => [
-        'text_total' => '总共 {0} 项',
-        'text_empty' => '空列表',
-        'filtered' => '{0} / {1}',
-        'filter_clear' => '显示全部',
+    'all_methods_if_empty'  => '为空默认为所有方法',
+    'all'                   => '全部',
+    'current_page'          => '当前页',
+    'selected_rows'         => '选择的行',
+    'upload'                => '上传',
+    'new_folder'            => '新建文件夹',
+    'time'                  => '时间',
+    'size'                  => '大小',
+    'between_start'         => '起始',
+    'between_end'           => '结束',
+    'next_page'             => '下一页',
+    'prev_page'             => '上一页',
+    'next_step'             => '下一步',
+    'prev_step'             => '上一步',
+    'done'                  => '完成',
+    'listbox'               => [
+        'text_total'         => '总共 {0} 项',
+        'text_empty'         => '空列表',
+        'filtered'           => '{0} / {1}',
+        'filter_clear'       => '显示全部',
         'filter_placeholder' => '过滤',
     ],
-    'responsive' => [
+    'responsive'      => [
         'display_all' => '显示全部',
-        'display' => '字段',
-        'focus' => '聚焦',
+        'display'     => '字段',
+        'focus'       => '聚焦',
     ],
     'uploader' => [
-        'add_new_media' => '添加文件',
-        'drag_file' => '或将文件拖到这里',
-        'max_file_limit' => 'The :attribute may not be greater than :max.',
-        'exceed_size' => '文件大小超出',
-        'interrupt' => '上传暂停',
-        'upload_failed' => '上传失败，请重试',
-        'selected_files' => '选中:num个文件，共:size。',
-        'selected_has_failed' => '已成功上传:success个文件，:fail个文件上传失败，<a class="retry"  href="javascript:"";">重新上传</a>失败文件或<a class="ignore" href="javascript:"";">忽略</a>',
-        'selected_success' => '共:num个(:size)，已上传:success个。',
-        'dot' => '，',
-        'failed_num' => '失败:fail个。',
-        'pause_upload' => '暂停上传',
-        'go_on_upload' => '继续上传',
-        'start_upload' => '开始上传',
+        'add_new_media'          => '添加文件',
+        'drag_file'              => '或将文件拖到这里',
+        'max_file_limit'         => 'The :attribute may not be greater than :max.',
+        'exceed_size'            => '文件大小超出',
+        'interrupt'              => '上传暂停',
+        'upload_failed'          => '上传失败，请重试',
+        'selected_files'         => '选中:num个文件，共:size。',
+        'selected_has_failed'    => '已成功上传:success个文件，:fail个文件上传失败，<a class="retry"  href="javascript:"";">重新上传</a>失败文件或<a class="ignore" href="javascript:"";">忽略</a>',
+        'selected_success'       => '共:num个(:size)，已上传:success个。',
+        'dot'                    => '，',
+        'failed_num'             => '失败:fail个。',
+        'pause_upload'           => '暂停上传',
+        'go_on_upload'           => '继续上传',
+        'start_upload'           => '开始上传',
         'upload_success_message' => '已成功上传:success个文件',
-        'go_on_add' => '继续添加',
-        'Q_TYPE_DENIED' => '对不起，不允许上传此类型文件',
-        'Q_EXCEED_NUM_LIMIT' => '对不起，已超出文件上传数量限制，最多只能上传:num个文件',
-        'F_EXCEED_SIZE' => '对不起，当前选择的文件过大',
-        'Q_EXCEED_SIZE_LIMIT' => '对不起，已超出文件大小限制',
-        'F_DUPLICATE' => '文件重复',
-        'confirm_delete_file' => '您确定要删除这个文件吗?',
+        'go_on_add'              => '继续添加',
+        'Q_TYPE_DENIED'          => '对不起，不允许上传此类型文件',
+        'Q_EXCEED_NUM_LIMIT'     => '对不起，已超出文件上传数量限制，最多只能上传:num个文件',
+        'F_EXCEED_SIZE'          => '对不起，当前选择的文件过大',
+        'Q_EXCEED_SIZE_LIMIT'    => '对不起，已超出文件大小限制',
+        'F_DUPLICATE'            => '文件重复',
+        'confirm_delete_file'    => '您确定要删除这个文件吗?',
     ],
-    'import_extension_confirm' => '确认导入拓展？',
-    'quick_create' => '快速创建',
-    'grid_items_selected' => '已选择 {n} 项',
-    'nothing_updated' => '没有任何数据被更改',
-    'welcome_back' => '欢迎回来，请登录您的账号。',
-    'documentation' => '文档',
-    'demo' => '示例',
-    'extensions' => '扩展',
-    'version' => '版本',
-    'current_version' => '当前版本',
-    'latest_version' => '最新版本',
-    'upgrade_to_version' => '更新至版本 :version',
-    'enable' => '启用',
-    'disable' => '禁用',
-    'uninstall' => '卸载',
-    'confirm_uninstall' => '您确定要卸载当前扩展吗？此操作将会移除扩展数据！',
-    'marketplace' => '应用市场',
-    'theme' => '主题',
-    'application' => '应用',
-    'install_from_local' => '本地安装',
-    'install_succeeded' => '安装成功',
+    'import_extension_confirm'  => '确认导入拓展？',
+    'quick_create'              => '快速创建',
+    'grid_items_selected'       => '已选择 {n} 项',
+    'nothing_updated'           => '没有任何数据被更改',
+    'welcome_back'              => '欢迎回来，请登录您的账号。',
+    'documentation'             => '文档',
+    'demo'                      => '示例',
+    'extensions'                => '扩展',
+    'version'                   => '版本',
+    'current_version'           => '当前版本',
+    'latest_version'            => '最新版本',
+    'upgrade_to_version'        => '更新至版本 :version',
+    'enable'                    => '启用',
+    'disable'                   => '禁用',
+    'uninstall'                 => '卸载',
+    'confirm_uninstall'         => '您确定要卸载当前扩展吗？此操作将会移除扩展数据！',
+    'marketplace'               => '应用市场',
+    'theme'                     => '主题',
+    'application'               => '应用',
+    'install_from_local'        => '本地安装',
+    'install_succeeded'         => '安装成功',
     'invalid_extension_package' => '安装包异常',
-    'copied' => '已复制',
-    'auth_failed' => '账号或密码错误',
-    'validation' => [
-        'match' => '与 :attribute 不匹配。',
+    'copied'                    => '已复制',
+    'auth_failed'               => '账号或密码错误',
+    'validation'               => [
+        'match'     => '与 :attribute 不匹配。',
         'minlength' => ':attribute 字符长度不能少于 :min。',
         'maxlength' => ':attribute 字符长度不能超出 :max。',
     ],
