@@ -5,17 +5,17 @@ return [
         'Extensions' => '扩展',
     ],
     'fields' => [
-        'alias' => '别名',
+        'alias'       => '别名',
         'description' => '描述',
-        'authors' => '开发者',
-        'homepage' => '主页',
-        'require' => '依赖',
+        'authors'     => '开发者',
+        'homepage'    => '主页',
+        'require'     => '依赖',
         'require_dev' => '开发环境依赖',
-        'name' => '包名',
-        'version' => '版本',
-        'enable' => '启用',
-        'config' => '配置',
-        'imported' => '导入',
+        'name'        => '包名',
+        'version'     => '版本',
+        'enable'      => '启用',
+        'config'      => '配置',
+        'imported'    => '导入',
     ],
     'options' => [
     ],
