@@ -2,226 +2,226 @@
 
 return [
     'scaffold' => [
-        'header' => '代碼生成器',
-        'choose' => '選擇已有數據表',
-        'table' => '表名',
-        'model' => '模型',
-        'controller' => '控制器',
-        'repository' => '數據倉庫',
-        'add_field' => '添加欄位',
-        'pk' => '主鍵',
-        'soft_delete' => '軟刪除',
-        'create_migration' => '創建表遷移文件',
-        'create_model' => '創建模型',
+        'header'            => '代碼生成器',
+        'choose'            => '選擇已有數據表',
+        'table'             => '表名',
+        'model'             => '模型',
+        'controller'        => '控制器',
+        'repository'        => '數據倉庫',
+        'add_field'         => '添加欄位',
+        'pk'                => '主鍵',
+        'soft_delete'       => '軟刪除',
+        'create_migration'  => '創建表遷移文件',
+        'create_model'      => '創建模型',
         'create_repository' => '創建數據倉庫',
         'create_controller' => '創建控制器',
-        'run_migrate' => '創建數據表',
-        'create_lang' => '創建翻譯文件',
-        'field' => '欄位',
-        'translation' => '翻譯',
-        'comment' => '註解',
-        'default' => '預設值',
-        'field_name' => '欄位名',
-        'type' => '類型',
-        'nullable' => '允許null',
-        'key' => '索引',
-        'translate_title' => '翻譯標題',
+        'run_migrate'       => '創建數據表',
+        'create_lang'       => '創建翻譯文件',
+        'field'             => '欄位',
+        'translation'       => '翻譯',
+        'comment'           => '註解',
+        'default'           => '預設值',
+        'field_name'        => '欄位名',
+        'type'              => '類型',
+        'nullable'          => '允許null',
+        'key'               => '索引',
+        'translate_title'   => '翻譯標題',
         'sync_translation_with_comment' => '同步翻譯與註釋',
     ],
     'client' => [
-        'delete_confirm' => '確認刪除?',
-        'confirm' => '確認',
-        'cancel' => '取消',
+        'delete_confirm'    => '確認刪除?',
+        'confirm'           => '確認',
+        'cancel'            => '取消',
         'refresh_succeeded' => '刷新成功!',
-        'close' => '關閉',
-        'submit' => '提交',
-        'selected_options' => '已選中:num個選項',
-        'exceed_max_item' => '已超出最大可選數量',
-        'no_preview' => '預覽失敗',
+        'close'             => '關閉',
+        'submit'            => '提交',
+        'selected_options'  => '已選中:num個選項',
+        'exceed_max_item'   => '已超出最大可選數量',
+        'no_preview'        => '預覽失敗',
 
         '500' => '系統繁忙，請稍後再試！',
         '403' => '對不起，您無訪問權限，請聯繫管理員。',
         '401' => '請先登入！',
         '419' => '對不起，當前頁面已失效，請刷新瀏覽器。',
     ],
-    'home' => '首頁',
-    'online' => '在線',
-    'login' => '登入',
-    'logout' => '登出',
-    'setting' => '設置',
-    'name' => '名稱',
-    'username' => '用戶名',
-    'old_password' => '舊密碼',
-    'password' => '密碼',
+    'home'                  => '首頁',
+    'online'                => '在線',
+    'login'                 => '登入',
+    'logout'                => '登出',
+    'setting'               => '設置',
+    'name'                  => '名稱',
+    'username'              => '用戶名',
+    'old_password'          => '舊密碼',
+    'password'              => '密碼',
     'password_confirmation' => '確認密碼',
-    'old_password_error' => '舊密碼輸入錯誤',
-    'remember_me' => '記住我',
-    'user_setting' => '用戶設置',
-    'avatar' => '頭像',
-    'list' => '列表',
-    'new' => '新增',
-    'create' => '創建',
-    'delete' => '刪除',
-    'remove' => '移除',
-    'edit' => '編輯',
-    'quick_edit' => '快速編輯',
-    'continue_editing' => '繼續編輯',
-    'continue_creating' => '繼續新增',
-    'view' => '查看',
-    'detail' => '詳細',
-    'browse' => '瀏覽',
-    'reset' => '重置',
-    'export' => '匯出',
-    'batch_delete' => '批次刪除',
-    'save' => '儲存',
-    'refresh' => '重新整理',
-    'order' => '排序',
-    'expand' => '展開',
-    'collapse' => '收起',
-    'filter' => '篩選',
-    'search' => '搜索',
-    'close' => '關閉',
-    'show' => '詳情',
-    'entries' => '條',
-    'captcha' => '驗證碼',
-    'action' => '操作',
-    'title' => '標題',
-    'description' => '簡介',
-    'back' => '返回',
-    'back_to_list' => '返回列表',
-    'submit' => '送出',
-    'menu' => '目錄',
-    'input' => '輸入',
-    'succeeded' => '成功',
-    'failed' => '失敗',
-    'delete_confirm' => '確認刪除？',
-    'delete_succeeded' => '刪除成功！',
-    'delete_failed' => '刪除失敗！',
-    'update_succeeded' => '更新成功！',
-    'update_failed' => '更新失敗 !',
-    'save_succeeded' => '儲存成功！',
-    'save_failed' => '儲存失敗 !',
-    'refresh_succeeded' => '刷新成功！',
-    'login_successful' => '登入成功！',
-    'choose' => '選擇',
-    'choose_file' => '選擇檔案',
-    'choose_image' => '選擇圖片',
-    'more' => '更多',
-    'deny' => '權限不足',
-    'administrator' => '管理員',
-    'roles' => '角色',
-    'permissions' => '權限',
-    'slug' => '標誌',
-    'created_at' => '建立時間',
-    'updated_at' => '更新時間',
-    'alert' => '警告',
-    'parent_id' => '父級',
-    'icon' => '圖示',
-    'uri' => '路徑',
-    'operation_log' => '操作記錄',
-    'parent_select_error' => '父級選擇錯誤',
-    'default' => '預設',
-    'table' => '表格',
-    'no_data' => '暫無數據',
-    'routes' => '路由',
-    'alias' => '別名',
-    'route_action' => '處理器',
-    'middleware' => '中間件',
-    'import' => '導入',
-    'is_not_import' => '未導入',
-    'selected_options' => '已選中:num個選項',
-    'method' => '方法',
-    'user' => '用戶',
-    'pagination' => [
+    'old_password_error'    => '舊密碼輸入錯誤',
+    'remember_me'           => '記住我',
+    'user_setting'          => '用戶設置',
+    'avatar'                => '頭像',
+    'list'                  => '列表',
+    'new'                   => '新增',
+    'create'                => '創建',
+    'delete'                => '刪除',
+    'remove'                => '移除',
+    'edit'                  => '編輯',
+    'quick_edit'            => '快速編輯',
+    'continue_editing'      => '繼續編輯',
+    'continue_creating'     => '繼續新增',
+    'view'                  => '查看',
+    'detail'                => '詳細',
+    'browse'                => '瀏覽',
+    'reset'                 => '重置',
+    'export'                => '匯出',
+    'batch_delete'          => '批次刪除',
+    'save'                  => '儲存',
+    'refresh'               => '重新整理',
+    'order'                 => '排序',
+    'expand'                => '展開',
+    'collapse'              => '收起',
+    'filter'                => '篩選',
+    'search'                => '搜索',
+    'close'                 => '關閉',
+    'show'                  => '詳情',
+    'entries'               => '條',
+    'captcha'               => '驗證碼',
+    'action'                => '操作',
+    'title'                 => '標題',
+    'description'           => '簡介',
+    'back'                  => '返回',
+    'back_to_list'          => '返回列表',
+    'submit'                => '送出',
+    'menu'                  => '目錄',
+    'input'                 => '輸入',
+    'succeeded'             => '成功',
+    'failed'                => '失敗',
+    'delete_confirm'        => '確認刪除？',
+    'delete_succeeded'      => '刪除成功！',
+    'delete_failed'         => '刪除失敗！',
+    'update_succeeded'      => '更新成功！',
+    'update_failed'         => '更新失敗 !',
+    'save_succeeded'        => '儲存成功！',
+    'save_failed'           => '儲存失敗 !',
+    'refresh_succeeded'     => '刷新成功！',
+    'login_successful'      => '登入成功！',
+    'choose'                => '選擇',
+    'choose_file'           => '選擇檔案',
+    'choose_image'          => '選擇圖片',
+    'more'                  => '更多',
+    'deny'                  => '權限不足',
+    'administrator'         => '管理員',
+    'roles'                 => '角色',
+    'permissions'           => '權限',
+    'slug'                  => '標誌',
+    'created_at'            => '建立時間',
+    'updated_at'            => '更新時間',
+    'alert'                 => '警告',
+    'parent_id'             => '父級',
+    'icon'                  => '圖示',
+    'uri'                   => '路徑',
+    'operation_log'         => '操作記錄',
+    'parent_select_error'   => '父級選擇錯誤',
+    'default'               => '預設',
+    'table'                 => '表格',
+    'no_data'               => '暫無數據',
+    'routes'                => '路由',
+    'alias'                 => '別名',
+    'route_action'          => '處理器',
+    'middleware'            => '中間件',
+    'import'                => '導入',
+    'is_not_import'         => '未導入',
+    'selected_options'      => '已選中:num個選項',
+    'method'                => '方法',
+    'user'                  => '用戶',
+    'pagination'            => [
         'range' => '從 :first 到 :last ，總共 :total 條',
     ],
-    'role' => '角色',
-    'permission' => '權限',
-    'route' => '路由',
-    'confirm' => '確認',
-    'cancel' => '取消',
-    'selectall' => '全選',
-    'http' => [
+    'role'                  => '角色',
+    'permission'            => '權限',
+    'route'                 => '路由',
+    'confirm'               => '確認',
+    'cancel'                => '取消',
+    'selectall'             => '全選',
+    'http'                  => [
         'method' => 'HTTP方法',
-        'path' => 'HTTP路徑',
+        'path'   => 'HTTP路徑',
     ],
-    'all_methods_if_empty' => '所有方法預設為空',
-    'all' => '全部',
-    'current_page' => '現在頁面',
-    'selected_rows' => '選擇的行',
-    'upload' => '上傳',
-    'new_folder' => '新建資料夾',
-    'time' => '時間',
-    'size' => '大小',
-    'between_start' => '起始',
-    'between_end' => '結束',
-    'next_page' => '下一頁',
-    'prev_page' => '上一頁',
-    'next_step' => '下一步',
-    'prev_step' => '上一步',
-    'done' => '完成',
-    'listbox' => [
-        'text_total' => '總共 {0} 項',
-        'text_empty' => '空列表',
-        'filtered' => '{0} / {1}',
-        'filter_clear' => '顯示全部',
+    'all_methods_if_empty'  => '所有方法預設為空',
+    'all'                   => '全部',
+    'current_page'          => '現在頁面',
+    'selected_rows'         => '選擇的行',
+    'upload'                => '上傳',
+    'new_folder'            => '新建資料夾',
+    'time'                  => '時間',
+    'size'                  => '大小',
+    'between_start'         => '起始',
+    'between_end'           => '結束',
+    'next_page'             => '下一頁',
+    'prev_page'             => '上一頁',
+    'next_step'             => '下一步',
+    'prev_step'             => '上一步',
+    'done'                  => '完成',
+    'listbox'               => [
+        'text_total'         => '總共 {0} 項',
+        'text_empty'         => '空列表',
+        'filtered'           => '{0} / {1}',
+        'filter_clear'       => '顯示全部',
         'filter_placeholder' => '過濾',
     ],
-    'responsive' => [
+    'responsive'      => [
         'display_all' => '顯示全部',
-        'display' => '欄位',
-        'focus' => '聚焦',
+        'display'     => '欄位',
+        'focus'       => '聚焦',
     ],
     'uploader' => [
-        'add_new_media' => '添加檔案',
-        'drag_file' => '或將檔案拖移到這裡',
-        'max_file_limit' => ':attribute不能超過:max.',
-        'exceed_size' => '檔案大小過大',
-        'interrupt' => '暫停上傳',
-        'upload_failed' => '上傳失敗，請再試一次',
-        'selected_files' => '選中:num個檔案，共:size。',
-        'selected_has_failed' => '已成功上傳:success個檔案，:fail個檔案上傳失敗，<a class="retry"  href="javascript:"";">重新上傳</a>失敗檔案或<a class="ignore" href="javascript:"";">忽略</a>',
-        'selected_success' => '共:num个(:size)，已上傳:success個。',
-        'dot' => '，',
-        'failed_num' => '失敗:fail個。',
-        'pause_upload' => '暫停上傳',
-        'go_on_upload' => '繼續上傳',
-        'start_upload' => '開始上傳',
+        'add_new_media'          => '添加檔案',
+        'drag_file'              => '或將檔案拖移到這裡',
+        'max_file_limit'         => ':attribute不能超過:max.',
+        'exceed_size'            => '檔案大小過大',
+        'interrupt'              => '暫停上傳',
+        'upload_failed'          => '上傳失敗，請再試一次',
+        'selected_files'         => '選中:num個檔案，共:size。',
+        'selected_has_failed'    => '已成功上傳:success個檔案，:fail個檔案上傳失敗，<a class="retry"  href="javascript:"";">重新上傳</a>失敗檔案或<a class="ignore" href="javascript:"";">忽略</a>',
+        'selected_success'       => '共:num个(:size)，已上傳:success個。',
+        'dot'                    => '，',
+        'failed_num'             => '失敗:fail個。',
+        'pause_upload'           => '暫停上傳',
+        'go_on_upload'           => '繼續上傳',
+        'start_upload'           => '開始上傳',
         'upload_success_message' => '已成功上傳:success個檔案',
-        'go_on_add' => '繼續添加',
-        'Q_TYPE_DENIED' => '對不起，無法上傳該類型檔案',
-        'Q_EXCEED_NUM_LIMIT' => '对不起，超出上傳數量限制，最多只能上傳:num個檔案',
-        'F_EXCEED_SIZE' => '对不起，當前選擇的檔案過大',
-        'Q_EXCEED_SIZE_LIMIT' => '对不起，檔案大小超出限制',
-        'F_DUPLICATE' => '檔案重複',
-        'confirm_delete_file' => '您確定要刪除該檔案嗎?',
+        'go_on_add'              => '繼續添加',
+        'Q_TYPE_DENIED'          => '對不起，無法上傳該類型檔案',
+        'Q_EXCEED_NUM_LIMIT'     => '对不起，超出上傳數量限制，最多只能上傳:num個檔案',
+        'F_EXCEED_SIZE'          => '对不起，當前選擇的檔案過大',
+        'Q_EXCEED_SIZE_LIMIT'    => '对不起，檔案大小超出限制',
+        'F_DUPLICATE'            => '檔案重複',
+        'confirm_delete_file'    => '您確定要刪除該檔案嗎?',
     ],
-    'import_extension_confirm' => '確認導入擴展?',
-    'quick_create' => '快速新增',
-    'grid_items_selected' => '已選擇 {n} 項',
-    'nothing_updated' => '沒有任何資料被更改',
-    'welcome_back' => '歡迎回來，請登入帳號',
-    'documentation' => '文檔',
-    'demo' => '範例',
-    'extensions' => '擴展',
-    'version' => '版本',
-    'current_version' => '当前版本',
-    'latest_version' => '當前版本',
-    'upgrade_to_version' => '更新至版本 :version',
-    'enable' => '啟用',
-    'disable' => '禁用',
-    'uninstall' => '卸載',
-    'confirm_uninstall' => '您確定要卸載當前擴展嗎？此操作將會移除擴展數據！',
-    'marketplace' => '應用市場',
-    'theme' => '主題',
-    'application' => '應用',
-    'install_from_local' => '本地安装',
-    'install_succeeded' => '本地安裝',
+    'import_extension_confirm'  => '確認導入擴展?',
+    'quick_create'              => '快速新增',
+    'grid_items_selected'       => '已選擇 {n} 項',
+    'nothing_updated'           => '沒有任何資料被更改',
+    'welcome_back'              => '歡迎回來，請登入帳號',
+    'documentation'             => '文檔',
+    'demo'                      => '範例',
+    'extensions'                => '擴展',
+    'version'                   => '版本',
+    'current_version'           => '当前版本',
+    'latest_version'            => '當前版本',
+    'upgrade_to_version'        => '更新至版本 :version',
+    'enable'                    => '啟用',
+    'disable'                   => '禁用',
+    'uninstall'                 => '卸載',
+    'confirm_uninstall'         => '您確定要卸載當前擴展嗎？此操作將會移除擴展數據！',
+    'marketplace'               => '應用市場',
+    'theme'                     => '主題',
+    'application'               => '應用',
+    'install_from_local'        => '本地安装',
+    'install_succeeded'         => '本地安裝',
     'invalid_extension_package' => '安裝包異常',
-    'copied' => '已複製',
-    'auth_failed' => '帳號或密碼錯誤',
-    'validation' => [
-        'match' => '與 :attribute 不匹配。',
+    'copied'                    => '已複製',
+    'auth_failed'               => '帳號或密碼錯誤',
+    'validation'               => [
+        'match'     => '與 :attribute 不匹配。',
         'minlength' => ':attribute 長度不能少於 :min。',
         'maxlength' => ':attribute 長度不能超過 :max。',
     ],
