@@ -5,17 +5,17 @@ return [
         'Extensions' => '擴展',
     ],
     'fields' => [
-        'alias' => '別名',
+        'alias'       => '別名',
         'description' => '說明',
-        'authors' => '作者',
-        'homepage' => '首頁',
-        'require' => '依賴',
+        'authors'     => '作者',
+        'homepage'    => '首頁',
+        'require'     => '依賴',
         'require_dev' => '開發環境依賴',
-        'name' => '名稱',
-        'version' => '版本',
-        'enable' => '啟用',
-        'config' => '設定',
-        'imported' => '導入',
+        'name'        => '名稱',
+        'version'     => '版本',
+        'enable'      => '啟用',
+        'config'      => '設定',
+        'imported'    => '導入',
     ],
     'options' => [
     ],
